requirements.txt
# Core dependencies
langchain-core>=0.1.0
langchain-chroma>=0.0.2
langchain-huggingface>=0.0.2
langchain-google-genai>=0.0.3
langchain-docling>=0.0.1
docling>=0.4.3
transformers>=4.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Vector store
chromadb>=0.4.18

# API and web dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.25.0
python-multipart>=0.0.6

# For PDF processing
pathlib>=1.0.1
